レビュー反映: テスト分割と冗長nullチェック除去
DragDeleteSelectionTest からカテゴリーテストと共有Fakeを別ファイルへ分割し
1ファイル200行制限を遵守。IsCategoryCompatible の冗長な null チェックを除去
（呼び出し側で保証済み・null方針準拠）。SchemaWatcher再生成分も取り込み。
