+++ moorestech_client/Assets/Scripts/Client.Tests/UIState/DragDeleteSelectionCategoryTest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d6fa761ae0c364672aeea4b10795815f
