+++ moorestech_client/Assets/Scripts/Client.Tests/UIState/FakeDeleteTarget.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c5a88b0ed84744a45a4e7449ce7bc8cb
